// src/app/games/interfaces/games-response.interface.ts
// Generated by https://quicktype.io

export interface GamesResponse {
    message:    string;
    statusCode: number;
    games?:     Game[];   // respuesta cuando se obtienen todos los juegos
    game?:      Game;     // respuesta cuando se crea un juego nuevo
}

export interface Game {
    _id?:         string;
    name:        string;
    description: string;
    imageURL:    string;
    createdAt?:   string;
    __v?:         number;
}